73d2266 BUG: Remove the GetTileScale logic.
0ecbbd0 Merge branch 'master' into charts_in_tile_display_resolve
6e53639 KWSys Nightly Date Stamp
e411a81 Merge topic 'sgconnectivity-warnings1' into master
59c57b3 COMP: Fix type, var, loop warnings in vtkStructuredGridConnectivity.
8304cd0 Merge branch 'pv-master'
3a4ef83 Merge branch 'pv-master'
9d9dcc8 KWSys Nightly Date Stamp
bf0e2ac Merge topic 'c++11-fixes' into master
0570314 KWSys Nightly Date Stamp
2b4ce3f KWSys Nightly Date Stamp
4209069 KWSys Nightly Date Stamp
7d13391 Merge topic 'bug-vtkcommand-eventid-wrapping' into master
bee52a2 BUG: Re-enable python wrapping of vtkCommand event ids.
802c4f7 Merge topic 'ConsistencyInDistanceNames' into master
221c74e Whitespace cleanup
...

+++ Graphics/Testing/Cxx/TestNamedComponents.cxx
@@ -49,7 +49,7 @@ int TestNamedComponents(int , char *[])
   points->SetNumberOfPoints(numPoints);
   for(i=0;i<numPoints;i++)
     {
-    double loc[3] = {i, i*i, 0};
+    double loc[3] = {static_cast<double>(i), static_cast<double>(i*i), 0.0};
     points->InsertPoint(i, loc);
     pointCoords->InsertTuple(i,loc);
     }
